baai-node/requirements.txt
lighthouseweb3>=0.1.0
cryptography>=3.4.7
python-dotenv>=0.19.0
aiohttp>=3.8.1 